+++ semver-gradle-plugin/test/kotlin/PropertiesTest.kt
@@ -1,0 +1,23 @@
+package com.javiersc.semver.gradle.plugin
+
+import kotlin.test.Test
+
+class PropertiesTest {
+
+    @Test
+    fun `gradle properties`() {
+        val projects =
+            getResource("properties")
+                .walkTopDown()
+                .mapNotNull { if (it.name == "settings.gradle.kts") it.parentFile else null }
+                .map { it.relativeTo(getResource("properties").parentFile).path }
+
+        projects.forEach {
+            testSandbox(
+                sandboxPath = it,
+                beforeTest = { generateInitialCommitAddVersionTagAndAddNewCommit() },
+                test = ::testSemVer,
+            )
+        }
+    }
+}